weight: allow one decimal place
- Measurement.gewicht_g int -> float (rounded to 1 decimal on write)
- SQLite stores decimals in the existing INTEGER-affinity column, no migration
- frontend: decimal input (step 0.1), German number formatting (comma) in
  table, sidebar and chart tooltip

+++ backend/app/models.py
@@ -28,7 +28,7 @@ class Measurement(SQLModel, table=True):
     id: Optional[int] = Field(default=None, primary_key=True)
     tortoise_id: int = Field(foreign_key="tortoise.id", index=True)
     datum: date
-    gewicht_g: Optional[int] = None
+    gewicht_g: Optional[float] = None  # eine Nachkommastelle erlaubt
     panzerlaenge_mm: Optional[int] = None
     notiz: Optional[str] = None
